+++ CHANGELOG.md
@@ -4,6 +4,12 @@ https://github.com/jcorporation/myGPIOd
 
 ***
 
+## myGPIOd v0.3.1 (2023-12-13)
+
+This release fixes a bug in the Alpine Linux package.
+
+***
+
 ## myGPIOd v0.3.0 (2023-10-09)
 
 This release adds bias support and requires therefor gpiod >= 1.5.0.